Move class FileCollection into Foundation\Kernel namespace
