+++ packages/framework/src/Foundation/Concerns/BaseFoundationCollection.php
@@ -11,7 +11,7 @@ use Illuminate\Support\Collection;
 /**
  * Base class for the kernel auto-discovery collections.
  *
- * @see \Hyde\Foundation\FileCollection
+ * @see \Hyde\Foundation\Kernel\FileCollection
  * @see \Hyde\Foundation\PageCollection
  * @see \Hyde\Foundation\RouteCollection
  * @see \Hyde\Framework\Testing\Unit\BaseFoundationCollectionTest
