+++ packages/framework/src/Foundation/Concerns/HandlesFoundationCollections.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Foundation\Concerns;
 
-use Hyde\Foundation\FileCollection;
+use Hyde\Foundation\Kernel\FileCollection;
 use Hyde\Foundation\PageCollection;
 use Hyde\Foundation\RouteCollection;
